Add SDL abstraction layer for graphics
This layer has been added to simplify the id and wl
code interactions with SDL. The plan is to have this codebase
be compatible with both SDL1 and SDL2, so performing most
of the SDL related stuff via an abstraction will make that easier
to support. I also plan on providing a similar abstraction for
the SDL input code.